pfloat: widen serde validate_normal pad product to u64 (pf-enj4)
validate_normal, the serde deserialize trust boundary, computed the
mantissa padding width as (limbs as u32) * 64 - precision. For a
precision near u32::MAX the product limbs * 64 reaches 2^32 and overflows
u32 before the subtraction brings it back into 0..64, panicking in any
overflow-checked build instead of returning the serde error the module
promises (ADR-0068: malformed input is rejected, never coerced or
panicked). The reachable trigger is a ~512 MB malformed payload, so
severity is low, but a panic at the trust boundary still breaks the
contract. The product now computes in u64; pad is provably below 64, so
the final u32 cast cannot truncate.

Found by adversarial review.

Verified: serde round-trip + rejection lane green (11 tests); clippy
full-union clean.

+++ src/serde_impls.rs
@@ -204,8 +204,11 @@ fn validate_normal(mantissa: &[u64], precision: u32) -> Result<(), &'static str>
         return Err("pfloat: Normal mantissa is not normalized (top bit clear)");
     }
     // Bits below the precision live in the least-significant limb; the
-    // padding width is `limbs * 64 - precision`, always in `0..64`.
-    let pad = (limbs as u32) * 64 - precision;
+    // padding width is `limbs * 64 - precision`, always in `0..64`. The
+    // product is computed in u64 so a near-`u32::MAX` precision (where
+    // `limbs * 64` reaches 2^32) cannot overflow before the subtraction
+    // brings it back into range; this site takes attacker bytes.
+    let pad = ((limbs as u64) * 64 - u64::from(precision)) as u32;
     if pad > 0 && mantissa[0] & ((1u64 << pad) - 1) != 0 {
         return Err("pfloat: Normal mantissa has nonzero bits below the precision");
     }
